fix: dev: clean up stderr debugging
Added helper function stderr_debug() for the simple debug calls
that check if debugging is enabled then do a single-line print.
It adds a prefix ($COMPONENT) to the message, to make debugging
flow easier to understand in multi-thread / multi-process apps.

The debug messages that are printed piece-wise while iterating
through a list are not handled by the new helper, so are mostly
left alone, aside from adding prefix of component name to the
message.