json.hh: exception for constructing from invalid text
The intention for json class is that wrapped json-c objects will
never be null, but this was able to be violated by the constructor
that has text string as argument. When the text string was not valid
JSON, the obj pointer turned out NULL.

Throw an exception instead, so this parsing failure case will need to be
handled right away instead of causing bugs later from a NULL pointer.